+++ HorseFluteAnywhere/release-notes.md
@@ -2,6 +2,7 @@
 
 # Release notes
 ## Upcoming release
+* When configuring the mod through Generic Mod Config Menu, you can now set [multi-key bindings](https://stardewvalleywiki.com/Modding:Player_Guide/Key_Bindings).
 * Fixed compatibility with [unofficial 64-bit mode](https://stardewvalleywiki.com/Modding:Migrate_to_64-bit_on_Windows).
 
 ## 1.1.3